# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/category": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new category to the system",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Add data Category",
                "parameters": [
                    {
                        "description": "Category Input",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/products.CategoryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category successfully added to the system",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid request or incomplete category data",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorBadRequest"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
...
        "/admin/product": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new Product to the system",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Add data Product",
                "parameters": [
                    {
                        "description": "Product Input",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/products.ProductInput"
                        }
                    },
                    {
                        "enum": [
                            "Man",
                            "Woman",
                            "Unisex"
                        ],
                        "type": "string",
                        "description": "Category Gender (Man, Woman, Unisex)",
                        "name": "categoryGender",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully added to the system",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid request or incomplete product data",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorBadRequest"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
...
        "/category": {
            "get": {
                "description": "Get detailed information of all data category and product based on category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get All Data Category",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of categories with details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Category"
                            }
                        }
                    },
                    "404": {
                        "description": "No categories found",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/product": {
            "get": {
                "description": "This endpoint is used to retrieve a list of all products along with detailed information for each product. It also allows you to get the details of a specific product by its ID. The response will include information like product name, description, price, and other relevant details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Retrieve All Product Data",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of products and their details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/products.ResponseProduct"
                            }
                        }
                    },
                    "404": {
                        "description": "Products not found",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/product/female": {
            "get": {
                "description": "Get detailed information of all data Product Female",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get All Data Product Female",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of female products with details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/products.ResponseProduct"
                            }
                        }
                    },
                    "404": {
                        "description": "No female products found",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
        "/product/male": {
            "get": {
                "description": "Get detailed information of all data Product Male",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get All Data Product Male",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of male products with details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/products.ResponseProduct"
                            }
                        }
                    },
                    "404": {
                        "description": "No male products found",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/products.ResponseErrorInternalServer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Category": {
            "type": "object",
            "required": [
                "category"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "maxLength": 100
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "products.CategoryInput": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                }
            }
        },
        "products.ProductInput": {
            "type": "object",
            "properties": {
                "categoryGender": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "dateRelase": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productName": {
                    "type": "string"
                },
                "typeProduct": {
                    "type": "string"
                }
            }
        },
        "products.ResponseErrorBadRequest": {
            "type": "object",
            "properties": {
                "errorFields": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "products.ResponseErrorInternalServer": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "products.ResponseErrorNotFound": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "products.ResponseProduct": {
            "type": "object",
            "properties": {
                "categoryGender": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "hargaProduct": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "namaProduct": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "products.ResponseSuccess": {
            "type": "object",
            "properties": {
                "Success": {
                    "type": "string"
                }
            }
        }
    }
}